1、完善用户组模块 2、完善用户模块

+++ gc-modules/gc-module-system/src/main/java/com/gc/system/controller/SysUserGroupController.java
@@ -2,9 +2,12 @@ package com.gc.system.controller;
 
 import com.gc.common.auth.model.SysUserPO;
 import com.gc.common.auth.utils.AuthUtils;
+import com.gc.common.base.http.HttpStatus;
 import com.gc.common.base.message.Result;
 import com.gc.starter.crud.controller.BaseController;
 import com.gc.starter.crud.query.PageQueryParameter;
+import com.gc.starter.log.annotation.Log;
+import com.gc.starter.log.constants.LogType;
 import com.gc.system.model.SysUserGroupPO;
 import com.gc.system.pojo.dto.UserGroupUserSaveDTO;
 import com.gc.system.pojo.dto.UserUserGroupSaveDTO;
@@ -35,6 +38,7 @@ public class SysUserGroupController extends BaseController<SysUserGroupService, 
     @Override
     @ApiOperation(value = "添加修改用户组")
     @PostMapping("saveUpdate")
+    @Log(value = "添加修改用户组", type = LogType.UPDATE)
     @PreAuthorize("hasPermission('sys:userGroup', 'save') or hasPermission('sys:userGroup', 'update')")
     public Result<Boolean> saveUpdate(@RequestBody @Valid SysUserGroupPO model) {
         return Result.success(this.service.saveOrUpdateWithAllUser(model, AuthUtils.getCurrentUserId()));
@@ -42,13 +46,17 @@ public class SysUserGroupController extends BaseController<SysUserGroupService, 
 
     /**
      * 执行删除
-     * @param idList
-     * @return
+     * @param idList 用户组ID集合
+     * @return 是否删除成功
      */
     @PreAuthorize("hasPermission('sys:userGroup', 'delete')")
     @ApiOperation(value = "通过ID批量删除用户组")
     @PostMapping("batchDeleteById")
+    @Log(value = "通过ID批量删除用户组", type = LogType.DELETE)
     public Result<Boolean> batchDeleteById(@RequestBody List<Long> idList) {
+        if (idList.isEmpty()) {
+            return Result.ofStatus(HttpStatus.PARAM_NOT_NULL, "用户组ID集合不能为空");
+        }
         return Result.success(this.service.removeByIds(idList));
     }
 
@@ -61,9 +69,9 @@ public class SysUserGroupController extends BaseController<SysUserGroupService, 
     }
 
     /**
-     * 通过ID查询用户信息
-     * @param id
-     * @return
+     * 通过ID查询用户集合
+     * @param id 用户组ID
+     * @return 用户集合
      */
     @PostMapping("listUserById")
     @ApiOperation(value = "通过用户组ID查询用户列表", httpMethod = "POST")
@@ -78,8 +86,8 @@ public class SysUserGroupController extends BaseController<SysUserGroupService, 
 
     /**
      * 通过ID查询用户ID集合
-     * @param id
-     * @return
+     * @param id 用户组ID
+     * @return 用户ID集合
      */
     @PostMapping(value = "listUserIdById")
     @ApiOperation(value = "通过用户组ID查询用户ID列表")
@@ -94,24 +102,26 @@ public class SysUserGroupController extends BaseController<SysUserGroupService, 
 
     /**
      * 保存用户
-     * @param parameter
-     * @return
+     * @param parameter 参数
+     * @return 是否保存成功
      */
     @PostMapping("saveUserGroupByGroupId")
     @PreAuthorize("hasPermission('sys:userGroup', 'saveUser')")
     @ApiOperation(value = "设置用户组包含的用户")
+    @Log(value = "设置用户组包含的用户", type = LogType.UPDATE)
     public Result<Boolean> saveUserGroupByGroupId(@RequestBody @Valid UserGroupUserSaveDTO parameter) {
         return Result.success(this.service.saveUserGroupByGroupId(parameter));
     }
 
     /**
      * 保存用户
-     * @param parameter
-     * @return
+     * @param parameter 参数
+     * @return 是否保存成功
      */
     @PostMapping("saveUserGroupByUserId")
     @ApiOperation(value = "设置用户所属用户组")
     @PreAuthorize("hasPermission('sys:userGroup', 'saveUser')")
+    @Log(value = "设置用户所属用户组", type = LogType.UPDATE)
     public Result<Boolean> saveUserGroupByUserId(@RequestBody @Valid UserUserGroupSaveDTO parameter) {
         return Result.success(this.service.saveUserGroupByUserId(parameter));
     }
